Add files via upload
Automation Excel to Power point though Microsoft Graph API. 

Takes a Mapping file stored in SharePoint that works like an index and the python code takes the NameRange in excel, copy that and paste it in a specific power point slice. 
Uses App Registration to connect to Microsoft Graph and GET the mapping data via API. 

The goal of this process was to create a report and improve the process that was manual and took around 8 hours and with the automations takes around 7 minutes for 200 slides. The time could vary depending the size of the excel files.